requirements.txt
django~=3.2.0
django-environ~=0.8.1
djangorestframework~=3.12.4
environ~=1.0
